+++ rust/src/control_flow/cfg_analysis_engine/contour_id.rs
@@ -2,7 +2,7 @@ use crate::control_flow::block_id::{ClassID, FunctionID};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub enum ContourID {
-    Module(i64),
+    Module(usize),
     Function(FunctionID),
     Class(ClassID),
 }
